Adjustment to new version

+++ Sage/Philosophy Path/Kerachole.js
@@ -6,35 +6,37 @@ const onUseItem = async () => {
   }
 
   const spawnKerachole = async (amountNoulith = 1) => {
-    const effectData = game.dfreds.effectInterface.findEffectByName('Kerachole').data.toObject();
+    const effectData = game.dfreds.effectInterface.findEffect({ effectName: 'Kerachole' }).toObject();
     effectData.description = effectData.description.replace("1d4", `${amountNoulith}d4`);
     effectData.changes = effectData.changes.map(change => {
-      if (change.key !== "flags.midi-qol.DR.all") return change;
-      return { ...change, value: `${amountNoulith}d4` }
+      if (change.key !== "system.traits.dm.midi.all") return change;
+      return { ...change, value: `- (floor(random() * ${amountNoulith * 4}) + 1)` };
     })
 
-    const prev = await game.canvas.tokens.ownedTokens.find((token) => token.name === 'Kerachole')
+    const spawnedKerachole = item.system.summons.summonedCreatures[0];
+    if (!spawnedKerachole) return;
+
+    const spawnedToken = spawnedKerachole.token;
+
+    const prev = await game.canvas.tokens.ownedTokens.find((token) => token.name === 'Kerachole' && token.id !== spawnedToken.id);
     if (prev) {
       await prev.document.delete();
     }
-
-    const [spawnedCreature2] = await warpgate.spawn("Kerachole", { token: { alpha: 1 } });
 
     const spawning = new Sequence()
       .wait(200)
       .effect()
       .file("jb2a.energy_field.02.above.blue")
-      .attachTo(spawnedCreature2)
+      .attachTo(spawnedToken)
       .scale(2.4)
       .persist()
-      .name(`Kerachole-${spawnedCreature2}`)
+      .name(`Kerachole-${spawnedToken.id}`)
       .fadeIn(300)
-      .fadeOut(300)
+      .fadeOut(300);
 
     await spawning.play();
 
-    const spawnedToken = game.canvas.tokens.get(spawnedCreature2);
-    game.dfreds.effectInterface.addEffectWith({ effectData, uuid: spawnedToken.actor.uuid });
+    game.dfreds.effectInterface.addEffect({ effectData, uuid: spawnedKerachole.uuid });
 
     const keracholeData = []
     for (let index = 0; index < amountNoulith; index++) {
@@ -42,16 +44,16 @@ const onUseItem = async () => {
       targets[index].document.delete();
     }
 
-    spawnedToken.actor.update({
+    spawnedKerachole.update({
       "system.attributes.hp.max": keracholeData.reduce((prev, crr) => prev + crr.hp, 0),
       "system.attributes.hp.value": keracholeData.reduce((prev, crr) => prev + crr.hp, 0)
     })
-    spawnedToken.actor.setFlag('world', 'Kerachole', keracholeData);
-    Hotbar.toggleDocumentSheet(actor.uuid)
+    spawnedKerachole.setFlag('world', 'Kerachole', keracholeData);
+    Hotbar.toggleDocumentSheet(actor.uuid);
   }
 
   const maxNoulithsToSend = Math.max(Math.min(item.actor.system.abilities.int.mod, 4), 1);
-  await spawnKerachole(Math.min(targets.length, maxNoulithsToSend))
+  await spawnKerachole(Math.min(targets.length, maxNoulithsToSend));
 }
 
 const onEffectEnd = async (args) => {
@@ -78,33 +80,32 @@ const onReacall = async () => {
 
   const spawnRecalledNouliths = async (recalleds = [{ hp: 0 }], avarageLostHp = 0) => {
     const actorToken = await game.canvas.tokens.ownedTokens.find((token) => token.actor.type === "character");
-    const gridOffset = game.canvas.grid.w * 0.5;
-    const noulithToGoPositions = [
-      { x: actorToken.x, y: actorToken.y },
-      { x: actorToken.x + gridOffset, y: actorToken.y },
-      { x: actorToken.x, y: actorToken.y + gridOffset },
-      { x: actorToken.x + gridOffset, y: actorToken.y + gridOffset }
-    ]
 
     recalleds.forEach(async (recalled, recalledIndex) => {
-      const [spawnedCreature] = await warpgate.spawnAt(noulithToGoPositions[recalledIndex], "Noulith", { token: { alpha: 0, elevation: 1 } });
+      const portal = new Portal().setLocation(actorToken.center).size(10);
+      portal.addCreature("Noulith", {
+        count: 1,
+        updateData: {
+          token: { elevation: 1, alpha: 0 },
+          actor: {
+            system: { attributes: { hp: { value: recalled.hp - avarageLostHp } } }
+          }
+        }
+      });
+      const [summon] = await portal.spawn();
+
       const spawning = new Sequence()
         .wait(200)
+        .effect()
+        .file("jb2a.swirling_sparkles.01.blue")
+        .atLocation(summon)
+        .wait(900)
         .animation()
-        .on(spawnedCreature)
+        .on(summon)
         .opacity(1.0)
 
       await spawning.play();
-
-      let token = canvas.tokens.placeables.find(t => t.id === spawnedCreature);
-      token.actor.update({ "system.attributes.hp.value": recalled.hp - avarageLostHp })
     })
-
-    new Sequence()
-      .effect()
-      .file("jb2a.swirling_sparkles.01.blue")
-      .atLocation(actorToken.id)
-      .play()
   }
 
   const recallNoulights = async (amount = 1) => {
@@ -118,11 +119,11 @@ const onReacall = async () => {
       const keracholeToken = await game.canvas.tokens.ownedTokens.find((token) => token.name === 'Kerachole')
       keracholeToken.document.delete();
     } else {
-      const effectData = game.dfreds.effectInterface.findEffectByName('Kerachole').data.toObject();
+      const effectData = game.dfreds.effectInterface.findEffect({ effectName: 'Kerachole' }).toObject();
       effectData.description = effectData.description.replace("1d4", `${amountNoulith}d4`);
       effectData.changes = effectData.changes.map(change => {
-        if (change.key !== "flags.midi-qol.DR.all") return change;
-        return { ...change, value: `${amountNoulith}d4` }
+        if (change.key !== "system.traits.dm.midi.all") return change;
+        return { ...change, value: `- (floor(random() * ${amountNoulith * 4}) + 1)` };
       })
 
       const recalleds = keracholeData.slice(0, amount)
@@ -131,7 +132,7 @@ const onReacall = async () => {
 
       keracholeData.splice(0, amount)
       const reducedMaxLife = recalleds.reduce((prev, crr) => prev + crr.hp, 0);
-      game.dfreds.effectInterface.addEffectWith({ effectData, uuid: actor.uuid });
+      game.dfreds.effectInterface.addEffect({ effectData, uuid: actor.uuid });
       actor.update({
         "system.attributes.hp.max": actor.system.attributes.hp.max - reducedMaxLife,
         "system.attributes.hp.value": (Math.min(actor.system.attributes.hp.max - reducedMaxLife, actor.system.attributes.hp.value) - (keracholeData.length * avarageLostHp))
